Split out CasterAuth into its own module
The CasterAuth code was turning out to be self-contained so factored it
out into its own module.

Termcast uses the first person to use that name and optional password to
"register" that name. Hash the password so subsequent comparisons will
use the same length. I doubt this is constant time (there's a
conditional in it!) but I hope it's close. Need to add tests next.