Use RingBuffer
The ring buffer is a naive method to store the recent buffer from a
caster. When a new watcher begins watching, the buffer is dumped to that
watcher to sync them. The buffer is also used at the beginning in order
to cache the login until all of the bytes have been sent.

The size of 90,000 is because that's what I remember the buffer size to
be. It's barely below a multiple of 4096, 90,112.