Test auth failures
This is kind of ugly. Rather than a blocking read with no timeout, a
read with a timeout should be used instead, or even something that
handles this better altogether (unsure at the moment).

May have to use mio in the tests as well. That will be tedious extra
code but may make other things easier later.