Disable linemode too
Disabling linemode and local echo has every key press sent immediately
to the server. This mimics the functionality of the original termcast.

Doing this causes the telnet client to send a fair (~40 bytes) of data
negotiating options. Really should factor out this into a separate class
or library but oh well! Instead going to hope for the best.